% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evm.R
\name{cv}
\alias{cv}
\title{Cost Variance (CV).}
\usage{
cv(ev, ac)
}
\arguments{
\item{ev}{Earned Value.}

\item{ac}{Actual Cost.}
}
\value{
The function returns the Cost Variance (CV) of work completed.
}
\description{
Cost Variance (CV).
}
\examples{
bac <- 100000
actual_per_complete <- 0.35
ev <- ev(bac, actual_per_complete)
actual_costs <- c(9000, 18000, 36000, 70000, 100000)
time_period <- 3
ac <- ac(actual_costs, time_period)

cv <- cv(ev, ac)
cat("Cost Variance (CV):", cv, "\n")
}
