% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evm.R
\name{ev}
\alias{ev}
\title{Earned Value (EV).}
\usage{
ev(bac, actual_per_complete)
}
\arguments{
\item{bac}{Budget at Completion (BAC) (total planned budget).}

\item{actual_per_complete}{Actual work completion percentage.}
}
\value{
The function returns the Earned Value (EV) of work completed.
}
\description{
Calculates the Earned Value (EV) of work completed based on the Budget at Completion
(BAC) and the actual work completion percentage.
}
\examples{
# Set the BAC and actual \% complete for a toy project.
bac <- 100000
actual_per_complete <- 0.35

# Calculate the EV and print the results.
ev <- ev(bac, actual_per_complete)
cat("Earned Value (EV):", ev, "\n")
}
\references{
Damnjanovic, Ivan, and Kenneth Reinschmidt. Data analytics for engineering and
construction project risk management. No. 172534. Cham, Switzerland: Springer, 2020.
}
\seealso{
\code{\link{pv}}, \code{\link{ac}}, \code{\link{sv}}, \code{\link{cv}},
\code{\link{spi}}, \code{\link{cpi}}, \code{\link{eac}}
}
