% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{format_help_overview}
\alias{format_help_overview}
\title{Format the /help overview}
\usage{
format_help_overview()

format_help_overview()
}
\value{
A markdown-formatted overview of all available /commands.
}
\description{
Format the /help overview

Format Help Overview of All Commands
}
\keyword{internal}
