% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_to_html}
\alias{plot_to_html}
\title{Generate a base64-encoded HTML img tag from a plot function}
\usage{
plot_to_html(plot_fn, width = 700, height = 450)
}
\description{
Generate a base64-encoded HTML img tag from a plot function
}
\keyword{internal}
