% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{pra_shiny_app}
\alias{pra_shiny_app}
\title{Create the PRA Shiny App Object}
\usage{
pra_shiny_app(model = "llama3.2", rag = TRUE, embed_model = "nomic-embed-text")
}
\arguments{
\item{model}{Ollama model name.}

\item{rag}{Whether to enable RAG.}

\item{embed_model}{Ollama embedding model.}
}
\value{
A Shiny app object.
}
\description{
Create the PRA Shiny App Object
}
\keyword{internal}
