% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsm.R
\name{print.dsm}
\alias{print.dsm}
\title{Print a DSM object.}
\usage{
\method{print}{dsm}(x, ...)
}
\arguments{
\item{x}{A \code{dsm} object returned by \code{\link[=parent_dsm]{parent_dsm()}} or \code{\link[=grandparent_dsm]{grandparent_dsm()}}.}

\item{...}{Additional arguments passed to \code{\link[=print.default]{print.default()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a DSM object.
}
