% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoidal.R
\name{print.nls}
\alias{print.nls}
\title{Print method for Sigmoidal Model}
\usage{
\method{print}{nls}(x, ...)
}
\arguments{
\item{x}{An object of class "nls" containing the fitted sigmoidal
model results.}

\item{...}{Additional arguments (not used).}
}
\description{
Displays the summary of the fitted sigmoidal model in a readable format.
}
