% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsm.R
\name{print.summary.dsm}
\alias{print.summary.dsm}
\title{Print a DSM summary.}
\usage{
\method{print}{summary.dsm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.dsm"} returned by \code{\link[=summary.dsm]{summary.dsm()}}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a DSM summary.
}
\examples{
S <- matrix(c(
  1, 1, 0,
  0, 1, 1
), nrow = 2, byrow = TRUE)
colnames(S) <- c("Design", "Build", "Test")
print(summary(parent_dsm(S)))
}
