% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcs.R
\name{print.summary.mcs}
\alias{print.summary.mcs}
\title{Print a Monte Carlo Simulation summary.}
\usage{
\method{print}{summary.mcs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.mcs"} returned by \code{\link[=summary.mcs]{summary.mcs()}}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Monte Carlo Simulation summary.
}
\examples{
task_dists <- list(
  list(type = "normal", mean = 10, sd = 2),
  list(type = "uniform", min = 8, max = 12)
)
print(summary(mcs(1000, task_dists)))
}
