% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoidal.R
\name{print.summary.pra_sigmoidal_fit}
\alias{print.summary.pra_sigmoidal_fit}
\title{Print a sigmoidal model fit summary.}
\usage{
\method{print}{summary.pra_sigmoidal_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.pra_sigmoidal_fit"} returned by
\code{\link[=summary.pra_sigmoidal_fit]{summary.pra_sigmoidal_fit()}}.}

\item{...}{Additional arguments passed to the \code{summary.nls} print method.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a sigmoidal model fit summary.
}
\examples{
data <- data.frame(time = 1:10, completion = c(
  5, 15, 40, 60, 70, 75, 80, 85, 90, 95
))
print(summary(fit_sigmoidal(data, "time", "completion", "logistic")))
}
