% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{print.summary.smm}
\alias{print.summary.smm}
\title{Print a Second Moment Method summary.}
\usage{
\method{print}{summary.smm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.smm"} returned by \code{\link[=summary.smm]{summary.smm()}}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Second Moment Method summary.
}
\examples{
print(summary(smm(c(10, 15, 20), c(4, 9, 16))))
}
