% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evm.R
\name{pv}
\alias{pv}
\title{Planned Value (PV).}
\usage{
pv(bac, schedule, time_period)
}
\arguments{
\item{bac}{Budget at Completion (BAC) (total planned budget).}

\item{schedule}{Vector of planned work completion (in terms of percentage) at each
time period.}

\item{time_period}{Current time period.}
}
\value{
The function returns the Planned Value (PV) of work completed.
}
\description{
Calculates the Planned Value (PV) of work completed based on the Budget at Completion
(BAC) and the planned schedule.
}
\examples{
# Set the BAC, schedule, and current time period for a toy project.
bac <- 100000
schedule <- c(0.1, 0.2, 0.4, 0.7, 1.0)
time_period <- 3

# Calculate the PV and print the results.
pv <- pv(bac, schedule, time_period)
cat("Planned Value (PV):", pv, "\n")

}
\references{
Damnjanovic, Ivan, and Kenneth Reinschmidt. Data analytics for engineering and
construction project risk management. No. 172534. Cham, Switzerland: Springer, 2020.
}
\seealso{
\code{\link{ev}}, \code{\link{ac}}, \code{\link{sv}}, \code{\link{cv}},
\code{\link{spi}}, \code{\link{cpi}}, \code{\link{eac}}
}
