% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{save_pra_plot}
\alias{save_pra_plot}
\title{Save a plot to a temp file and return path}
\usage{
save_pra_plot(plot_fn, name, width = 700, height = 450)
}
\description{
Save a plot to a temp file and return path
}
\keyword{internal}
