% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evm.R
\name{spi}
\alias{spi}
\title{Schedule Performance Index (SPI).}
\usage{
spi(ev, pv)
}
\arguments{
\item{ev}{Earned Value.}

\item{pv}{Planned Value.}
}
\value{
The function returns the Schedule Performance Index (SPI) of work completed.
}
\description{
Schedule Performance Index (SPI).
}
\examples{
bac <- 100000
schedule <- c(0.1, 0.2, 0.4, 0.7, 1.0)
time_period <- 3
pv <- pv(bac, schedule, time_period)
actual_per_complete <- 0.35
ev <- ev(bac, actual_per_complete)

spi <- spi(ev, pv)
cat("Schedule Performance Index (SPI):", spi, "\n")
}
