% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{validate_task_dists}
\alias{validate_task_dists}
\title{Validate task distributions parsed from JSON}
\usage{
validate_task_dists(task_dists, tool_name = "mcs_tool")
}
\description{
Validate task distributions parsed from JSON
}
\keyword{internal}
