% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evm.R
\name{ac}
\alias{ac}
\title{Actual Cost (AC).}
\usage{
ac(actual_costs, time_period)
}
\arguments{
\item{actual_costs}{Vector of actual costs incurred at each time period.}

\item{time_period}{Current time period.}
}
\value{
The function returns the Actual Cost (AC) of work completed.
}
\description{
Actual Cost (AC).
}
\examples{
actual_costs <- c(9000, 18000, 36000, 70000, 100000)
time_period <- 3

ac <- ac(actual_costs, time_period)
cat("Actual Cost (AC):", ac, "\n")
}
