% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{cost_pdf}
\alias{cost_pdf}
\title{Cost Probability Density.}
\usage{
cost_pdf(
  num_sims,
  risk_probs,
  means_given_risks,
  sds_given_risks,
  base_cost = 0
)
}
\arguments{
\item{num_sims}{Number of random samples to draw from the mixture model.}

\item{risk_probs}{A vector of probabilities for each risk event 'R_i'.}

\item{means_given_risks}{A vector of means of the normal distribution for cost 'A' given each risk event 'R_i'.}

\item{sds_given_risks}{A vector of standard deviations of the normal distribution for cost 'A' given each risk event 'R_i'.}

\item{base_cost}{The baseline cost given no risk event occurs.}
}
\value{
A numeric vector of random samples from the mixture model.
}
\description{
Bayesian inference for calculating random samples of a cost given risk event(s).
}
\examples{
# Example with three risk events
num_sims <- 1000
risk_probs <- c(0.3, 0.5, 0.2)
means_given_risks <- c(10000, 15000, 5000)
sds_given_risks <- c(2000, 1000, 1000)
base_cost <- 2000
samples <- cost_pdf(
  num_sims = num_sims,
  risk_probs = risk_probs,
  means_given_risks = means_given_risks,
  sds_given_risks = sds_given_risks,
  base_cost = base_cost
)
hist(samples, breaks = 30, col = "skyblue", main = "Histogram of Cost", xlab = "Cost")
}
