% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evm.R
\name{cv}
\alias{cv}
\title{Cost Variance (CV).}
\usage{
cv(ev, ac)
}
\arguments{
\item{ev}{Earned Value.}

\item{ac}{Actual Cost.}
}
\value{
The function returns the Cost Variance (CV) of work completed.
}
\description{
Calculates the Cost Variance (CV) of work completed based on the Earned Value (EV)
and Actual Cost (AC).
}
\examples{
# Set the BAC and actual \% complete for an example project.
bac <- 100000
actual_per_complete <- 0.35

# Calcualte the EV
ev <- ev(bac, actual_per_complete)

# Set the actual costs and current time period and calculate the AC.
actual_costs <- c(9000, 18000, 36000, 70000, 100000)
time_period <- 3
ac <- ac(actual_costs, time_period)

# Calculate the CV and print the results.
cv <- cv(ev, ac)
cat("Cost Variance (CV):", cv, "\n")
}
\references{
Damnjanovic, Ivan, and Kenneth Reinschmidt. Data analytics for engineering and
construction project risk management. No. 172534. Cham, Switzerland: Springer, 2020.
}
\seealso{
\code{\link{pv}}, \code{\link{ev}}, \code{\link{ac}}, \code{\link{sv}},
\code{\link{spi}}, \code{\link{cpi}}, \code{\link{eac}}
}
