% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evm.R
\name{eac}
\alias{eac}
\title{Estimate at Completion (EAC).}
\usage{
eac(bac, cpi)
}
\arguments{
\item{bac}{Budget at Completion (BAC) (total planned budget).}

\item{cpi}{Cost Performance Index (CPI), the efficiency of cost utilization.}
}
\value{
The function returns the Estimate at Completion (EAC).
}
\description{
Estimate at Completion (EAC).
}
\examples{
# Set the BAC and CPI for a toy project.
bac <- 100000
cpi <- 0.83

# Calculate the EAC
eac <- eac(bac, cpi)
print(eac)
}
