% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsm.R
\name{grandparent_dsm}
\alias{grandparent_dsm}
\title{Risk-based 'Grandparent' Design Structure Matrix (DSM).}
\usage{
grandparent_dsm(sm, rm)
}
\arguments{
\item{sm}{Resource-Task Matrix 'S' giving the links (arcs) between resources and tasks.}

\item{rm}{Risk-Resource Matrix 'R' giving the links (arcs) between risks and resources.}
}
\value{
The function returns the Risk-based 'Grandparent' DSM 'G' giving the number
of risks shared between each task.
}
\description{
Risk-based 'Grandparent' Design Structure Matrix (DSM).
}
\examples{
sm <- matrix(c(1, 1, 0, 0, 1, 0, 0, 1, 1), nrow = 3, ncol = 3)
rm <- matrix(c(1, 1, 1, 1, 0, 0), nrow = 2, ncol = 3)
cat("Resource-Task Matrix:\n")
print(sm)
cat("\nRisk-Resource Matrix:\n")
print(rm)
risk_dsm <- grandparent_dsm(sm, rm)
cat("\nRisk-based 'Grandparent' DSM:\n")
print(risk_dsm)
}
\seealso{
\url{https://ascelibrary.org/doi/abs/10.1061/AJRUA6.0001041}
}
