% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsm.R
\name{parent_dsm}
\alias{parent_dsm}
\title{Resource-based 'Parent' Design Structure Matrix (DSM).}
\usage{
parent_dsm(S)
}
\arguments{
\item{S}{Resource-Task Matrix 'S' giving the links (arcs) between resources and tasks.}
}
\value{
The function returns the Resource-based 'Parent' DSM 'P' giving the number
of resources shared between each task.
}
\description{
This function computes the Resource-based 'Parent' Design Structure Matrix (DSM)
from a given Resource-Task Matrix 'S'. The 'Parent' DSM indicates the number of resources
shared between each pair of tasks in a project.
}
\examples{
# Set the S matrix for a toy project and print the results.
s <- matrix(c(1, 1, 0, 0, 1, 0, 0, 1, 1), nrow = 3, ncol = 3)
cat("Resource-Task Matrix:\n")
print(s)

# Calculate the Resource-based Parent DSM and print the results.
resource_dsm <- parent_dsm(s)
cat("\nResource-based 'Parent' DSM:\n")
print(resource_dsm)

}
\references{
Govan, Paul, and Ivan Damnjanovic. "The resource-based view on project risk management."
Journal of construction engineering and management 142.9 (2016): 04016034.
}
