% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsm.R
\name{parent_dsm}
\alias{parent_dsm}
\title{Resource-based 'Parent' Design Structure Matrix (DSM).}
\usage{
parent_dsm(sm)
}
\arguments{
\item{sm}{Resource-Task Matrix 'S' giving the links (arcs) between resources and tasks.}
}
\value{
The function returns the Resource-based 'Parent' DSM 'P' giving the number
of resources shared between each task.
}
\description{
Resource-based 'Parent' Design Structure Matrix (DSM).
}
\examples{
sm <- matrix(c(1, 1, 0, 0, 1, 0, 0, 1, 1), nrow = 3, ncol = 3)
cat("Resource-Task Matrix:\n")
print(sm)
resource_dsm <- parent_dsm(sm)
cat("\nResource-based 'Parent' DSM:\n")
print(resource_dsm)
}
