% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRA-package.R
\docType{package}
\name{PRA-package}
\alias{PRA}
\alias{PRA-package}
\title{PRA: Project Risk Analysis}
\description{
Project Risk Analysis (PRA) provides functions for quantitative risk analysis
of engineering and construction projects. The package includes methods for
Monte Carlo simulation, earned value management, Bayesian inference, sensitivity
analysis, and sigmoidal growth curve fitting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paulgovan.github.io/PRA/}
  \item \url{https://github.com/paulgovan/PRA}
  \item Report bugs at \url{https://github.com/paulgovan/PRA/issues}
}

}
\author{
\strong{Maintainer}: Paul Govan \email{paul.govan2@gmail.com} (\href{https://orcid.org/0000-0002-1821-8492}{ORCID}) [copyright holder]

}
\keyword{internal}
