% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRA-package.R
\docType{package}
\name{PRA-package}
\alias{PRA}
\alias{PRA-package}
\title{PRA: Project Risk Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Data analysis for Project Risk Management via the Second Moment Method, Monte Carlo Simulation, Contingency Analysis, Sensitivity Analysis, Earned Value Management, Learning Curves, Bayesian Methods, Bayesian Networks, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paulgovan.github.io/PRA/}
  \item \url{https://github.com/paulgovan/PRA}
  \item Report bugs at \url{https://github.com/paulgovan/PRA/issues}
}

}
\author{
\strong{Maintainer}: Paul Govan \email{paul.govan2@gmail.com} (\href{https://orcid.org/0000-0002-1821-8492}{ORCID}) [copyright holder]

}
\keyword{internal}
