% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcs.R
\name{print.mcs}
\alias{print.mcs}
\title{Print method for Monte Carlo Simulation results.}
\usage{
\method{print}{mcs}(x, ...)
}
\arguments{
\item{x}{An object of class "mcs".}

\item{...}{Additional arguments (not used).}
}
\description{
Displays the total mean, variance, standard deviation, and percentiles of the
Monte Carlo Simulation results in a readable format.
}
