% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{prob_net_update}
\alias{prob_net_update}
\title{Update a Probabilistic Network of Project Risks.}
\usage{
prob_net_update(
  graph,
  add_links = NULL,
  remove_links = NULL,
  update_distributions = NULL
)
}
\arguments{
\item{graph}{An existing probabilistic network created by `prob_net()`.}

\item{add_links}{Optional. A data frame with columns `source` and `target` to add new links.}

\item{remove_links}{Optional. A data frame with columns `source` and `target` to remove existing links.}

\item{update_distributions}{Optional. A named list of distributions to update. Format follows `prob_net()`.}
}
\value{
An updated `prob_net` object with modified links and/or distributions.
}
\description{
This function updates an existing probabilistic network by adding or removing dependencies (edges)
and updating probability distributions for nodes.
}
