% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evm.R
\name{pv}
\alias{pv}
\title{Planned Value (PV).}
\usage{
pv(bac, schedule, time_period)
}
\arguments{
\item{bac}{Budget at Completion (total planned budget).}

\item{schedule}{Vector of planned work completion (in terms of percentage) at each time period.}

\item{time_period}{Current time period.}
}
\value{
The function returns the Planned Value (PV) of work completed.
}
\description{
Planned Value (PV).
}
\examples{
bac <- 100000
schedule <- c(0.1, 0.2, 0.4, 0.7, 1.0)
time_period <- 3
pv <- pv(bac, schedule, time_period)
cat("Planned Value (PV):", pv, "\n")
}
