% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{smm}
\alias{smm}
\title{Second Moment Analysis.}
\usage{
smm(mean, var, cov = NULL)
}
\arguments{
\item{mean}{The mean vector.}

\item{var}{The variance vector.}

\item{cov}{The covariance matrix.}
}
\value{
The function returns a list of the total mean, variance, and standard deviation.
}
\description{
Second Moment Analysis.
}
\examples{
mean <- c(10, 15, 20)
var <- c(4, 9, 16)
cov <- matrix(c(4, 1, 0.5, 1, 9, 2, 0.5, 2, 16), nrow = 3, byrow = TRUE)
result <- smm(mean, var, cov)
print(result)
}
